+++ src/Whoops/Exception/Frame.php
@@ -112,10 +112,10 @@ class Frame implements Serializable
     public function getFileContents()
     {
         if ($this->fileContentsCache === null && $filePath = $this->getFile()) {
-            // Leave the stage early when 'Unknown' is passed
+            // Leave the stage early when 'Unknown' or '[internal]' is passed
             // this would otherwise raise an exception when
             // open_basedir is enabled.
-            if ($filePath === "Unknown") {
+            if ($filePath === "Unknown" || $filePath === '[internal]') {
                 return null;
             }